server_code/requirements.txt
qrcode[pil]
pillow
opencv-python
